requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
Werkzeug~=3.0.6
python-dotenv~=1.0.1
Flask-JWT-Extended~=4.6.0
redis~=5.2.0
flask_sqlalchemy
flask-swagger-ui
